Fixs + upgrade to latest sqlalchemy, flask

+++ nano/models/billing.py
@@ -7,7 +7,7 @@ from nano.extensions import db
 class Plan(db.Model):
     __tablename__ = 'plan'
 
-    id          = db.Column(db.Integer(11), primary_key=True, nullable=False)
+    id          = db.Column(db.Integer, primary_key=True, nullable=False)
     name        = db.Column(db.Unicode(20))
     gateway_uid = db.Column(db.Unicode(255))
     description = db.Column(db.Unicode(255))
@@ -26,9 +26,9 @@ class Plan(db.Model):
 
 class Subscription(db.Model):
     __tablename__ = 'subscription'
-    id          = db.Column(db.Integer(11), primary_key=True, nullable=False)
-    user_id     = db.Column(db.Integer(11), db.ForeignKey('user.id'), nullable=False)
-    plan_id     = db.Column(db.Integer(11), db.ForeignKey('plan.id'), nullable=False)
+    id          = db.Column(db.Integer, primary_key=True, nullable=False)
+    user_id     = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    plan_id     = db.Column(db.Integer, db.ForeignKey('plan.id'), nullable=False)
     active      = db.Column(db.Boolean, nullable=False, default=False)
     start_date  = db.Column(db.DateTime, nullable=True)
     end_date    = db.Column(db.DateTime, nullable=True)
@@ -43,9 +43,9 @@ class Subscription(db.Model):
 class Transaction(db.Model):
     __tablename__ = 'transaction'
 
-    id              = db.Column(db.Integer(11), primary_key=True, nullable=False)
-    user_id         = db.Column(db.Integer(11), db.ForeignKey('user.id'), nullable=False)
-    subscription_id = db.Column(db.Integer(11), db.ForeignKey('subscription.id'), nullable=False)
+    id              = db.Column(db.Integer, primary_key=True, nullable=False)
+    user_id         = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    subscription_id = db.Column(db.Integer, db.ForeignKey('subscription.id'), nullable=False)
     success         = db.Column(db.Boolean, default=False)
     amount          = db.Column(db.Numeric(8, 2), default=0)
     charge_id       = db.Column(db.Unicode(100), nullable=True)
